Isolate TheHeroMoment section
