+++ components/landing/TheHeroMoment/VersionInfo.vue
